Split PokemonResponse off from PokemonSpecies
By separating the response sent to clients into it's own type, we can
have a stronger assurance that the correct data is sent back to the
client - ie: the response sent back isn't just a reserialised response
from the upstream pokeapi.
This also simplifies some of the getters on the PokemonSpecies struct,
as it no longer has to account for cases that should not ocurr, if the
design is adhered to correctly - ie: get_first_description would never
be called on a PokemonSpecies object transformed for return as a
response.
Also allows for the removal of unnecessary Clone derives from existing
models.